Refactor the github code into separate files
