+++ src/components/PrPanel.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { PrTable } from "./PrTable";
-import { searchForPr } from "../lib/github";
-import GithubQuery from "../lib/query";
+import { searchForPr } from "../lib/github/searchForPr";
+import GithubSearchQuery from "../lib/github/GithubSearchQuery";
 import { useGithubApi } from "../lib/useGithubApi";
 
 export const PrPanel: React.FC<{
@@ -10,7 +10,7 @@ export const PrPanel: React.FC<{
   columns: import("./PrTable/PrTable").TableCols;
 }> = ({ filter, repos, columns }) => {
   const query = [
-    new GithubQuery()
+    new GithubSearchQuery()
       .repo(...repos, { quote: true })
       .is("pr")
       .toQuery(),
