+++ src/components/PrTable/PrTable.js
@@ -13,7 +13,7 @@ import { PrRow } from "./PrRow";
 
 /**
  * @typedef TableProps
- * @prop {import('../../lib/github').PrForLink[]} prs
+ * @prop {import('../../lib/github/queries').PrForLink[]} prs
  * @prop {TableCols=} columns
  */
