+++ src/lib/github/prStatusCommit.js
@@ -1,0 +1,6 @@
+/**
+ * @param {import('./queries').PrWithStatus} pr
+ */
+export function prStatusCommit(pr) {
+  return pr.commits.nodes[0].commit;
+}
