+++ src/lib/github/repoFromUrl.js
@@ -1,0 +1,5 @@
+/**
+ * @param {string} url
+ */
+export const repoFromUrl = (url) =>
+  new URL(url).pathname.split("/").slice(1, 3);
