+++ src/lib/usePopperAlerter.js
@@ -1,4 +1,5 @@
 import { useOutsideAlerter } from "@aha-app/aha-develop-react";
+// @ts-ignore
 import { usePopper } from "https://cdn.skypack.dev/react-popper";
 import { useCallback, useRef, useState } from "react";
